building 세분화, Building ScriptableObject 세분화, 건물 확장,
공격력, 체력 모두 float로 변환

+++ ProjectBS/Assets/_BsScripts/Building/Buildings/AttackBuilding/AttackBuilding_Projectile.cs
@@ -12,7 +12,7 @@ public class AttackBuilding_Projectile : AttackBuildingBase
         set { AreaBuildingData = value; }
     }
 
-    [SerializeField] public short _atkPower;
+    [SerializeField] public float _atkPower;
     [SerializeField] public float _atkDelay;
     [SerializeField] public float _atkProjectileSize;
     [SerializeField] public float _atkProjectileSpeed; // 건물 투사체 속도
